+++ src/test/converter2/issues/gh2970.ts
@@ -1,0 +1,10 @@
+/** Comment */
+class Class {}
+
+/** Comment */
+const Var = 123;
+
+/** Comment */
+function Func() {}
+
+export type { Class, Func, Var };